requirements.txt
Django>=5.2.8,<6
psycopg2-binary>=2.9,<3
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.4,<6
